Change RandomSplits script to allow for incremental execution
The summary file is still only generated after everything has run, so
force-aborting the script and continuing running later will create an
inconsistent summary file.

The intended use is to iteratively extend the set of event logs it runs
on, or manually filter the instances list for smaller chunks to run.